Renamed ref-resolver back to uuid resolver
