+++ lib/Storage/UuidResolver.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace PhpBench\Storage;
+
+use PhpBench\Storage\UuidResolverInterface;
+
+/**
+ * Wrapper class for the UUID resolver interface which will always return the
+ * UUID
+ */
+final class UuidResolver
+{
+    /**
+     * @var UuidResolverInterface
+     */
+    private $innerResolver;
+
+    public function __construct(UuidResolverInterface $innerResolver)
+    {
+        $this->innerResolver = $innerResolver;
+    }
+
+    public function resolve(string $ref): string
+    {
+        return $this->innerResolver->resolve($ref) ?? $ref;
+    }
+}
