+++ tests/Unit/Storage/UuidResolver/LatestResolverTest.php
@@ -10,13 +10,13 @@
  *
  */
 
-namespace PhpBench\Tests\Unit\Storage\RefResolver;
+namespace PhpBench\Tests\Unit\Storage\UuidResolver;
 
 use PhpBench\Registry\Registry;
 use PhpBench\Storage\DriverInterface;
 use PhpBench\Storage\HistoryEntry;
 use PhpBench\Storage\HistoryIteratorInterface;
-use PhpBench\Storage\RefResolver\LatestResolver;
+use PhpBench\Storage\UuidResolver\LatestResolver;
 use PhpBench\Tests\TestCase;
 use RuntimeException;
